iiitb_server/text_summarisation/requirements.txt
python-dotenv
openai
langchain_groq
langchain_openai
langchain_core
fastapi
flask
opencv-python
aiortc
aiohttp
sentencepiece
langchain-community
langchain-huggingface
